+++ src/refactoring/App.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { CartPage } from "./components/cart/CartPage.tsx";
+import { CartPage } from "./pages/CartPage.tsx";
 import { AdminPage } from "./pages/AdminPage.tsx";
 import { useCoupons, useProducts } from "./hooks";
 import { initialProducts, initialCoupons } from "./shared/constants";
