+++ src/refactoring/pages/CartPage.tsx
@@ -1,10 +1,10 @@
-import { Coupon, Product } from "../../../types.ts";
-import { useCart } from "../../hooks/index.ts";
-import { OrderSummary } from "./order/OrderSummary.tsx";
-import { CouponSelector } from "./coupon/CouponSelector.tsx";
-import { CartList } from "./CartList";
-import { ProductList } from "./product/ProductList.tsx";
-import { CartItem } from "./CartItem.tsx";
+import { Coupon, Product } from "../../types.ts";
+import { useCart } from "../hooks/index.ts";
+import { OrderSummary } from "../components/cart/order/OrderSummary.tsx";
+import { CouponSelector } from "../components/cart/coupon/CouponSelector.tsx";
+import { CartList } from "../components/cart/CartList.tsx";
+import { ProductList } from "../components/cart/product/ProductList.tsx";
+import { CartItem } from "../components/cart/CartItem.tsx";
 
 interface Props {
   products: Product[];
